frr: change order type to u16
The order in a route map entry is an u16, not an u32 - adapt this
here to match the actual type inside FRR.

Link: https://lore.proxmox.com/20260505153720.412180-7-s.hanreich@proxmox.com

+++ proxmox-frr/src/ser/route_map.rs
@@ -120,7 +120,7 @@ impl RouteMapName {
 /// address or adding a metric, bgp community, or local preference.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RouteMapEntry {
-    pub seq: u32,
+    pub seq: u16,
     pub action: AccessAction,
     #[serde(default)]
     pub matches: Vec<RouteMapMatch>,
